requirements.txt
pdf2docx==0.5.8
pytesseract==0.3.10
python-docx==1.1.2
pdf2image==1.17.0
functions-framework==3.*
flask==2.*